update project and form behaviours

+++ src/services/project.ts
@@ -13,28 +13,29 @@ class ProjectService {
     }
   };
 
-  getProject = async (id: string) => {
-    const user = await Project.findById(id);
+  getProject = async (owner: string, id: string) => {
+    const project = await Project.findOne({ _id: id, owner });
 
-    if (!user) {
-      throw new Error();
+    if (!project) {
+      throw new Error('invalid-project');
     }
 
-    return user;
+    return project;
   };
 
-  getProjects = async (): Promise<IProjectDocument[]> => {
-    return await Project.find();
+  getProjects = async (owner: string): Promise<IProjectDocument[]> => {
+    return await Project.find({ owner });
   };
 
   updateProject = async (
+    owner: string,
     id: string,
     newData: Record<string, any>
   ): Promise<IProjectDocument> => {
     const updatables = ['name', 'description'];
 
     try {
-      const project = await Project.findById(id);
+      const project = await Project.findOne({ _id: id, owner });
 
       updatables.forEach((key) => {
         if (newData[key] != undefined) {
@@ -52,8 +53,11 @@ class ProjectService {
     }
   };
 
-  deleteProject = async (id: string): Promise<IProjectDocument> => {
-    const project = await Project.findOneAndDelete({ _id: id });
+  deleteProject = async (
+    owner: string,
+    id: string
+  ): Promise<IProjectDocument> => {
+    const project = await Project.findOneAndDelete({ _id: id, owner });
 
     if (!project) {
       throw new Error();
